add(precarga users y posts exitosa)

+++ api/src/models/User.js
@@ -2,17 +2,17 @@ const { DataTypes } = require('sequelize');
 
 module.exports = (sequelize) => {
     sequelize.define('user', {
-        id:{
+        id: {
             type: DataTypes.UUID,
             defaultValue: DataTypes.UUIDV4,
             allowNull: false,
-            primaryKey: true, 
+            primaryKey: true,
         },
         firstName: {
             type: DataTypes.STRING(255),
             allowNull: false,
         },
-        lastName:{
+        lastName: {
             type: DataTypes.STRING(255),
             allowNull: false,
         },
@@ -20,23 +20,23 @@ module.exports = (sequelize) => {
             type: DataTypes.INTEGER,
             allowNull: true,
         },
-        ID_Passport:{
+        ID_Passport: {
             type: DataTypes.STRING,
             allowNull: false,
         },
         address: {
-            type:DataTypes.STRING,
+            type: DataTypes.STRING,
             allowNull: false,
         },
         phoneNumber: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.BIGINT,
             allowNull: false,
         },
         email: {
             type: DataTypes.STRING,
             allowNull: false,
         },
-        summary:{
+        summary: {
             type: DataTypes.TEXT,
             allowNull: true,
         },
@@ -52,14 +52,14 @@ module.exports = (sequelize) => {
             type: DataTypes.INTEGER,
             allowNull: true,
         },
-        isVaccinated:{
+        isVaccinated: {
             type: DataTypes.BOOLEAN,
             allowNull: true,
         },
         isNew: {
             type: DataTypes.BOOLEAN,
             allowNull: true,
-        },        
+        },
         isAdmin: {
             type: DataTypes.BOOLEAN,
             allowNull: true
